+++ src/main/EnemyBossBullet.java
@@ -21,7 +21,7 @@ public class EnemyBossBullet extends GameObject{
 	}
 		
 	public Rectangle getBounds() {
-		return new Rectangle(x, y, 16, 16);
+		return new Rectangle((int)x, (int)y, 16, 16);
 	}
 	
 	public void tick() {
@@ -38,7 +38,7 @@ public class EnemyBossBullet extends GameObject{
 		
 		if(y >= Game.HEIGHT) handler.removeObject(this);
 		
-		handler.addObject(new Trail(x, y, ID.Trail, Color.red, 16, 16, 0.075f, handler));
+		handler.addObject(new Trail((int)x, (int)y, ID.Trail, Color.red, 16, 16, 0.075f, handler));
 		
 		
 		
@@ -52,7 +52,7 @@ public class EnemyBossBullet extends GameObject{
 		g2d.draw(getBounds());*/
 		
 		g.setColor(Color.red);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int)x, (int)y, 16, 16);
 	}
 
 }
